Code refactoring due to readability sake
