+++ test_conformance/extensions/cl_khr_semaphore/CMakeLists.txt
@@ -3,6 +3,7 @@ set(MODULE_NAME CL_KHR_SEMAPHORE)
 set(${MODULE_NAME}_SOURCES
          main.cpp
          test_semaphores.cpp
+         test_semaphores_queries.cpp
          test_semaphores_negative_release_retain.cpp
          test_semaphores_negative_getinfo.cpp
          test_semaphores_negative_wait.cpp
